requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pan-os-python==1.12.3
pan-python==0.17.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
urllib3==2.5.0
